// types/typechain/hardhat-diamond-abi/HardhatDiamondABI.sol/Kresko.ts
export type CollateralAssetStruct = {
  factor: FixedPoint.UnsignedStruct;
  oracle: PromiseOrValue<string>;
  kreskoAsset: PromiseOrValue<string>;
  decimals: PromiseOrValue<BigNumberish>;
  exists: PromiseOrValue<boolean>;
};
// ...
export declare namespace FixedPoint {
  export type UnsignedStruct = { rawValue: PromiseOrValue<BigNumberish> };

  export type UnsignedStructOutput = [BigNumber] & { rawValue: BigNumber };
}